chore: moving model to models folder
