+++ src/Nullinside.Api.TwitchBot.Tests/ChatRules/BestCheapViewersTests.cs
@@ -2,6 +2,7 @@ using Moq;
 
 using Nullinside.Api.Common.Twitch;
 using Nullinside.Api.TwitchBot.ChatRules;
+using Nullinside.Api.TwitchBot.Model;
 
 namespace Nullinside.Api.TwitchBot.Tests.ChatRules;
