+++ src/Nullinside.Api.TwitchBot/ChatRules/Nezhna.cs
@@ -1,8 +1,10 @@
 ﻿using Nullinside.Api.Common.Twitch;
 using Nullinside.Api.Model;
-using Nullinside.Api.Model.Ddl;
+using Nullinside.Api.TwitchBot.Model;
 
 using TwitchLib.Client.Models;
+
+using TwitchUserConfig = Nullinside.Api.Model.Ddl.TwitchUserConfig;
 
 namespace Nullinside.Api.TwitchBot.ChatRules;
